moving user & org TimezoneExtension code to shared class
